+++ frontend/app/src/main/java/com/team13/fooriend/ui/screen/search/SearchScreen.kt
@@ -1,0 +1,33 @@
+package com.team13.fooriend.ui.screen.search
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.TextField
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import com.team13.fooriend.ui.component.ScreenChangeBar
+
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun SearchPage(navController: NavController){
+    val (search, searchValue) = remember { mutableStateOf("") }
+    Column(
+        modifier = Modifier.fillMaxSize(),
+        verticalArrangement = Arrangement.Top,
+        horizontalAlignment = Alignment.CenterHorizontally,
+    ){
+        Spacer(modifier = Modifier.height(20.dp))
+        TextField(value = search, onValueChange = searchValue)
+    }
+    ScreenChangeBar(navController)
+}
